fix edit dialog
